feat: add push32 codegen
