+++ Cargo.toml
@@ -11,6 +11,7 @@ keywords = ["ethereum", "evm"]
 llvm-sys = "180.0.0"
 melior = { version = "0.18.1", features = ["ods-dialects"] }
 mlir-sys = "0.2.2"
+num-bigint = "0.4.5"
 thiserror = "1.0.57"
 
 [build-dependencies]
