+++ src/lib.rs
@@ -1,2 +1,5 @@
 pub mod codegen;
+pub mod context;
+pub mod errors;
+pub mod module;
 pub mod opcodes;
